revue waybar wofi

+++ link.sh
@@ -4,25 +4,30 @@
 DOTFILES_DIR=~/hypr-dotfiles
 CONFIG_DIR=~/.config
 
+FILES=("hypr" "kitty" "waybar" "fastfetch" "wofi" "cava")
 
-FILES=("hypr" "kitty" "waybar" "fastfetch" "wofi")
-
-echo "🔗 Linking dotfiles from $DOTFILES_DIR to $CONFIG_DIR..."
+echo "Linking dotfiles from $DOTFILES_DIR to $CONFIG_DIR..."
 
 for folder in "${FILES[@]}"; do
     src="$DOTFILES_DIR/$folder"
     dest="$CONFIG_DIR/$folder"
 
-    if [ -L "$dest" ]; then
-        echo "🔁 Removing existing symlink: $dest"
-        rm "$dest"
-    elif [ -d "$dest" ]; then
-        echo "🗑️ Removing existing directory: $dest"
-        rm -rf "$dest"
-    fi
+    # ➤ Corrigé : espace après [ et avant ]
+    if [ ! -e "$src" ]; then 
+        echo "Copying $dest to $src (doesn't exist in dotfiles)"
+        cp -r "$dest" "$DOTFILES_DIR"
+    else
+        if [ -L "$dest" ]; then
+            echo "Removing existing symlink: $dest"
+            rm "$dest"
+        elif [ -d "$dest" ]; then
+            echo "Removing existing directory: $dest"
+            rm -rf "$dest"
+        fi
 
-    echo "✅ Linking $src -> $dest"
-    ln -s "$src" "$dest"
+        echo "Linking $src -> $dest"
+        ln -s "$src" "$dest"
+    fi
 done
 
-echo "🎉 All done!"
+echo "All done!"
